Add roddy install and add err check
roddy install [tagid] will check your local repository for the existence
of a specific Roddy tag. If it exists, a temporary directory will be
created, Roddy will be cloned to it and the requested tag will be
compiled. Finally it will be copied to your local installation. Nothing
is changed in your repository!

Add check for compileplugin, if a Roddy jar exists. If not, abort.